style(agents): :art: Fix linting errors
I completely forgot about eslint but its okay because claude did it for me. Although it tossed //eslint-ignore at everything which is kinda real but I will need to deal with that at some point here. I am too tired now and want to commit so I can run this.